+++ AngularApp/projects/app-service-diagnostics/src/app/shared/shared.module.ts
@@ -126,6 +126,7 @@ import { FabMessageBarModule } from '@angular-react/fabric/lib/components/messag
 import { FabLinkModule } from '@angular-react/fabric/lib/components/link';
 import { FabPivotModule } from '@angular-react/fabric/lib/components/pivot';
 import { DownloadReportComponent } from './components/download-report/download-report.component';
+import { OptInsightsService } from './services/optinsights/optinsights.service';
 
 @NgModule({
     declarations: [
@@ -321,7 +322,8 @@ export class SharedModule {
                 GenericCommsService,
                 GroupByPipe,
                 LocalBackendService,
-                ABTestingService
+                ABTestingService,
+                OptInsightsService
             ]
         };
     }
