+++ AngularApp/projects/diagnostic-data/src/lib/diagnostic-data.module.ts
@@ -161,6 +161,7 @@ import { QueryResponseService } from './services/query-response.service';
 import { WorkflowConditionNodeComponent } from './components/workflow-condition-node/workflow-condition-node.component';
 import { OptInsightsEnablementComponent } from './components/opt-insights-enablement/opt-insights-enablement.component';
 import { OptInsightsMarkdownComponent } from './components/opt-insights-markdown/opt-insights-markdown.component';
+import { OptInsightsGenericService } from './services/optinsights.service';
 import { WorkflowAcceptUserinputComponent } from './components/workflow-accept-userinput/workflow-accept-userinput.component';
 import { VideoComponent } from './components/video/video.component';
 import { SafePipe } from './pipe/safe.pipe';
@@ -352,7 +353,8 @@ export class DiagnosticDataModule {
                 GenericOpenAIChatService,
                 OpenAIArmService,
                 ChatGPTContextService,
-                QueryResponseService
+                QueryResponseService,
+                OptInsightsGenericService
             ]
         };
     }
