+++ AngularApp/projects/diagnostic-data/src/lib/services/portal-action.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { OptInsightsResource, OptInsightsTimeContext } from '../models/optinsights';
 
 @Injectable({
   providedIn: 'root'
@@ -12,4 +13,10 @@ export class PortalActionGenericService {
   }
 
   public openFeedbackPanel(){}
+
+  public openOptInsightsBladewithTimeRange(appInsightsResourceUri: OptInsightsResource, optInsightsTimeContext: OptInsightsTimeContext) {
+  }
+
+  public openOptInsightsBlade(appInsightsResourceUri: OptInsightsResource) {}
+
 }
